# api/openapi.yaml
openapi: 3.0.3
info:
  title: Creatorr API
  version: 0.1.0
  description: |
    REST API for Creatorr (creator VOD daemon).
    HTMX HTML routes and SSE (`/api/events`) are out of this contract - see AGENTS.md.
servers:
  - url: /
    description: Local / Docker
...
paths:
...
  /api/roots:
    get:
      tags: [library]
      operationId: listRoots
      summary: List root folders
      responses:
        "200":
          description: Roots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RootFolder"
    post:
      tags: [library]
      operationId: createRoot
      summary: Create a root folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRootRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootFolder"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/roots/{id}:
    patch:
      tags: [library]
      operationId: updateRoot
      summary: Update a root folder
      parameters:
        - $ref: "#/components/parameters/RootId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRootRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootFolder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/quality-profiles:
    get:
      tags: [library]
      operationId: listQualityProfiles
      summary: List quality profiles
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QualityProfile"
    post:
      tags: [library]
      operationId: createQualityProfile
      summary: Create a quality profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQualityProfileRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityProfile"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/quality-profiles/{id}:
    patch:
      tags: [library]
      operationId: updateQualityProfile
      summary: Update a quality profile
      parameters:
        - $ref: "#/components/parameters/QualityProfileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQualityProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/series:
    get:
      tags: [library]
      operationId: listSeries
      summary: List series with sources and counts
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
    post:
      tags: [library]
      operationId: createSeries
      summary: Create a series (optional first source)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/series/{id}:
    get:
      tags: [library]
      operationId: getSeries
      summary: Series detail with sources and videos
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [library]
      operationId: deleteSeries
      summary: Delete a series and its sources/indexed entries (library files on disk are kept; UI delete always removes library files after confirm)
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/series/{id}/scan:
    post:
      tags: [library]
      operationId: scanSeries
      summary: Enqueue full scan for sources that still need indexing
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "201":
          description: Scan task enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueTaskResponse"
        "400":
          description: No eligible sources (full scan done or domains inactive)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Series not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Scan already queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/series/{id}/metadata-rescan:
    post:
      tags: [library]
      operationId: metadataRescanSeries
      summary: Enqueue metadata rescan for existing videos (no discovery)
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "201":
          description: Metadata rescan task enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueTaskResponse"
        "400":
          description: No monitored sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Series not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Metadata rescan already queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/series/{id}/sources:
    post:
      tags: [library]
      operationId: addSource
      summary: Add a source URL to a series
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Series not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/series/{id}/sources/{sourceId}:
    patch:
      tags: [library]
      operationId: updateSource
      summary: Update a source
      parameters:
        - $ref: "#/components/parameters/SeriesId"
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [library]
      operationId: deleteSource
      summary: Delete a source and all of its videos
      description: >
        Removes the source, hard-deletes every indexed video that belonged to it
        (including on-disk media/sidecars), and cancels related scan/download tasks.
        To keep videos, unmonitor the source instead.
      parameters:
        - $ref: "#/components/parameters/SeriesId"
        - $ref: "#/components/parameters/SourceId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/videos/{id}:
    get:
      tags: [library]
      operationId: getVideo
      summary: Get one indexed video
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/videos/{id}/want:
    post:
      tags: [library]
      operationId: wantVideo
      summary: Set status to wanted from ignored, deleted, or missing; does not enqueue download
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Status cannot be wanted again
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/videos/{id}/metadata-rescan:
    post:
      tags: [library]
      operationId: metadataRescanVideo
      summary: Enqueue metadata rescan for one indexed video
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "201":
          description: Metadata rescan task enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueTaskResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Metadata rescan already queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    RootId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    QualityProfileId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SeriesId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    VideoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code (e.g. CookieInvalid, DownloadFailed, RemuxFailed, PackFailed)
        message:
          type: string
          description: Short user-facing text
        detail:
          type: string
          description: Optional operator/debug detail
...
    EnqueueTaskResponse:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
...
    RootFolder:
      type: object
      required: [id, name, path]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        path:
          type: string
        retention_ttl_seconds:
          type: integer
          format: int64
          nullable: true
          description: Seconds after acquire before files may be removed. Empty = no retention.
    CreateRootRequest:
      type: object
      required: [path]
      properties:
        name:
          type: string
          description: Display label; empty uses the last path segment.
        path:
          type: string
          description: Absolute path on disk (relative paths are rejected).
        retention_ttl_seconds:
          type: integer
          format: int64
          nullable: true
    UpdateRootRequest:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        retention_ttl_seconds:
          type: integer
          format: int64
          nullable: true
    QualityProfile:
      type: object
      required: [id, name, format_selector, maturity_redownload_hours, maturity_sidecar_hours, sponsorblock_mark, sponsorblock_remove, sponsorblock_reencode_cut, sponsorblock_info_cards, verify_media]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        format_selector:
          type: string
          description: yt-dlp -f style format selector; Creatorr passes it to the domain handler as --format
        maturity_redownload_hours:
          type: integer
          minimum: 0
          maximum: 168
          description: Hours after upload_date for one blind media maturity pass (0 = off for all series on this profile).
        maturity_sidecar_hours:
          type: integer
          minimum: 0
          maximum: 8760
          description: Hours after upload_date for one NFO/thumb/subs refresh (0 = off). UI edits this as days. Does not rewrite info.json.
        sponsorblock_mark:
          type: array
          items:
            type: string
          description: SponsorBlock categories to embed as extra MKV chapters (explicit list; empty = off). Must be disjoint from sponsorblock_remove.
        sponsorblock_remove:
          type: array
          items:
            type: string
          description: SponsorBlock categories to cut/skip (explicit list; empty = off). Must be disjoint from sponsorblock_mark.
        sponsorblock_reencode_cut:
          type: boolean
          description: When remove is non-empty, re-encode keep segments for accurate cut timestamps (bitrate/codec-matched). Off = stream-copy keyframe snap.
        sponsorblock_info_cards:
          type: boolean
          description: When reencode_cut is on and remove is non-empty, insert hard-cut info cards at each skip. Rejected if set without reencode_cut.
        verify_media:
          type: boolean
          description: >
            When true, after archive pack enqueue a low-priority system media_verify (null-decode).
            With maturity_redownload_hours > 0, young first packs skip verify until the maturity re-download
            (or when maturity will never run). Default false.
    CreateQualityProfileRequest:
      type: object
      required: [name, format_selector]
      properties:
        name:
          type: string
        format_selector:
          type: string
          description: yt-dlp -f style format selector
        maturity_redownload_hours:
          type: integer
          minimum: 0
          maximum: 168
          description: Hours after upload_date for media maturity (0 = off; default 0)
        maturity_sidecar_hours:
          type: integer
          minimum: 0
          maximum: 8760
          description: Hours after upload_date for sidecar maturity (0 = off; default 0). UI edits as days.
        sponsorblock_mark:
          type: array
          items:
            type: string
        sponsorblock_remove:
          type: array
          items:
            type: string
        sponsorblock_reencode_cut:
          type: boolean
        sponsorblock_info_cards:
          type: boolean
        verify_media:
          type: boolean
          description: Post-pack media verify (default false)
    UpdateQualityProfileRequest:
      type: object
      properties:
        name:
          type: string
        format_selector:
          type: string
          description: yt-dlp -f style format selector
        maturity_redownload_hours:
          type: integer
          minimum: 0
          maximum: 168
        maturity_sidecar_hours:
          type: integer
          minimum: 0
          maximum: 8760
        sponsorblock_mark:
          type: array
          items:
            type: string
        sponsorblock_remove:
          type: array
          items:
            type: string
        sponsorblock_reencode_cut:
          type: boolean
        sponsorblock_info_cards:
          type: boolean
        verify_media:
          type: boolean
...
    Series:
      type: object
      required: [id, title, root_id, quality_profile_id, monitored, added_at]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        root_id:
          type: integer
          format: int64
        quality_profile_id:
          type: integer
          format: int64
        monitored:
          type: boolean
          description: Series-level monitored; never rewrites source/video/domain monitored flags
        auto_ignore_media_types:
          type: array
          items:
            type: string
          description: >
            yt-dlp media_type values to auto-ignore (e.g. short, livestream). Empty = none.
            Missing media_type is never auto-ignored. Applied at index create when type is known,
            at download via --match-filters, and at stream pack_stream from the urls extract
            (indexed, then marked ignored; no cache_beginning).
        added_at:
          type: string
          format: date-time
        root_name:
          type: string
        quality_profile_name:
          type: string
        video_count:
          type: integer
          format: int64
        downloaded_count:
          type: integer
          format: int64
        wanted_count:
          type: integer
          format: int64
        source_count:
          type: integer
          format: int64
        sources:
          type: array
          items:
            $ref: "#/components/schemas/Source"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
          description: Present on series detail GET only
...